requirements.txt
accelerate==1.7.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
blinker==1.9.0
branca==0.8.1
cachetools==5.5.2
certifi==2025.4.26
cffi==1.15.1
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
contourpy==1.3.0
cv-3==1.2.0
cvxopt==1.3.2
cycler==0.12.1
deprecation==2.1.0
exceptiongroup==1.3.0
fastapi==0.115.12
filelock==3.18.0
folium==0.19.6
fonttools==4.57.0
fsspec==2025.5.0
geopandas==1.0.1
gitdb==4.0.12
GitPython==3.1.44
graphviz==0.20.3
h11==0.16.0
httptools==0.6.4
huggingface-hub==0.31.4
idna==3.10
importlib_resources==6.5.2
intervaltree==3.1.0
Jinja2==3.1.6
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.7
lxml==5.4.0
MarkupSafe==3.0.2
matplotlib==3.9.4
meteostat==1.6.8
mip==1.15.0
mpmath==1.3.0
narwhals==1.39.1
networkx==3.2.1
numpy==2.0.2
opencv-python==4.11.0.86
osmnx==2.0.3
packaging==24.2
pandas==2.2.3
pillow==11.2.1
pm4py==2.7.15.1
protobuf==6.31.0
psutil==7.0.0
pyarrow==20.0.0
pycparser==2.22
pydantic==2.11.4
pydantic_core==2.33.2
pydeck==0.9.1
pydotplus==2.0.2
pyogrio==0.11.0
pyparsing==3.2.3
pyproj==3.6.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-slugify==8.0.4
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rpds-py==0.25.0
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.13.1
sentencepiece==0.2.0
shapely==2.0.7
six==1.17.0
slugify==0.0.1
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
starlette==0.46.2
streamlit==1.45.1
streamlit_folium==0.25.0
sympy==1.14.0
tenacity==9.1.2
text-unidecode==1.3
threadpoolctl==3.6.0
tokenizers==0.21.1
toml==0.10.2
tomli==2.2.1
torch==2.7.0
torchaudio==2.7.0
torchvision==0.22.0
tornado==6.5
tqdm==4.67.1
transformers==4.52.3
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.2
watchdog==6.0.0
watchfiles==1.0.5
websockets==15.0.1
xyzservices==2025.4.0
zipp==3.21.0

langchain~=0.3.26
pytest~=8.4.1
pytesseract~=0.3.13
reportlab~=4.4.2